cinema/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.1
django-debug-toolbar==3.5.0
django-modeltranslation==0.18.5
django-phonenumber-field==6.3.0
kombu==5.2.4
packaging==21.3
phonenumberslite==8.12.54
Pillow==9.3.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pyparsing==3.0.9
pytz==2022.5
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
